fix: preserve command palette coverage history
